Remove BpfBaseTest processgroup test.
In practice this has to pass already due to netd_updatable
init attaching BPF_CGROUP_INET_{E,IN}GRESS programs.

This allows removing the processgroup shared lib dependency,
which should allow switching it from C to C++ mangling.

Thus we might be able to revert the revert (ie. reapply):
  https://android-review.googlesource.com/c/platform/system/core/+/3366140
in
  https://android-review.googlesource.com/c/platform/system/core/+/3369297

Test: TreeHugger
Bug: 379796721
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I865d6705e4a6477d428261ef182173df1f5b4672